Refactor: Remove FeedCapture component and update Canvas styles location
- Removed the unused FeedCapture component from the project
- Relocated Canvas styles from feed_capture to microscope_feed for better coherence

#140
